Improved PowerUps
Added an array that stores information about which powerups are active and how long they will be active for. Added a visual countdown timer for powerups.  Added a parachute graphic to the powerups. Modified the isTouching() function to work with elements that are not direct children of the body element.